requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-anymail==8.5
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==21.2
django-compressor==3.1
django-crispy-forms==1.13.0
django-mpesa==2.0.10
django-sass-processor==1.1
django-static-precompiler==2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
libsass==0.21.0
mailchimp-marketing==3.0.74
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
